third_party/WebKit/LayoutTests/media/encrypted-media/encrypted-media-waiting-for-a-key.html

<!DOCTYPE html>
<html>
    <head>
        <title>Waiting for a key.</title>
        <script src="encrypted-media-utils.js"></script>
        <script src="../../resources/testharness.js"></script>
        <script src="../../resources/testharnessreport.js"></script>
    </head>
    <body>
        <video id="testVideo"></video>
        <div id="log"></div>
        <script>
            async_test(function(test)
            {
                var video = document.getElementById('testVideo');
                var mediaKeySession = null;
                var hasSessionUpdateSucceeded = false;
                var encryptedEventCount = 0;
                var waitingForKeyEvent = true;

                // As this code doesn't wait for the 'message' event to avoid
                // race conditions with 'waitingforkey', specify the key ID
                // used by the encrypted content.
                var keyId = stringToUint8Array('0123456789012345');
                var rawKey = new Uint8Array([0xeb, 0xdd, 0x62, 0xf1, 0x68, 0x14, 0xd2, 0x7b,
                                             0x68, 0xef, 0x12, 0x2a, 0xfc, 0xe4, 0xae, 0x3c]);

                function onEncrypted(event)
                {
                    assert_equals(event.target, video);
                    assert_true(event instanceof window.MediaEncryptedEvent);
                    assert_equals(event.type, 'encrypted');

                    // The same decryption key is used by both the audio and
                    // the video streams so only create a session once. Create
                    // the session on the second event to ensure that we
                    // see both events.
                    if (++encryptedEventCount != 2)
                        return;

                    mediaKeySession = video.mediaKeys.createSession();
                    mediaKeySession.generateRequest(event.initDataType, event.initData).catch(function(error) {
                        forceTestFailureFromPromise(test, error);
                    });
                }

                function onWaitingForKey(event)
                {
                    assert_equals(event.target, video);
                    assert_equals(event.type, 'waitingforkey');

                    // Make sure the video is NOT paused and not progressing
                    // before the session is updated.
                    assert_false(video.paused);
                    assert_equals(video.currentTime, 0);

                    // TODO(jrummell): waitingforkey event should only occur
                    // once. http://crbug.com/461903
//                  assert_true(waitingForKeyEvent, 'Multiple waitingforkey events');
                    waitingForKeyEvent = false;

                    // Add the key needed to decrypt.
                    updateSession();
                }

                function updateSession()
                {
                    var jwkSet = stringToUint8Array(createJWKSet(createJWK(keyId, rawKey)));
                    mediaKeySession.update(jwkSet).then(function(result) {
                        hasSessionUpdateSucceeded = true;
                    }).catch(function(error) {
                        forceTestFailureFromPromise(test, error);
                    });
                }

                function onPlaying(event)
                {
                    // Not using waitForEventAndRunStep() to avoid
                    // EVENT(onTimeUpdate) logs.
                    video.addEventListener('timeupdate', onTimeUpdate, true);
                }

                function onTimeUpdate(event)
                {
                    if (event.target.currentTime < 0.5 || !hasSessionUpdateSucceeded)
                        return;

                    test.done();
                }

                navigator.requestMediaKeySystemAccess('org.w3.clearkey', [{}]).then(function(access) {
                    return access.createMediaKeys();
                }).then(function(mediaKeys) {
                    waitForEventAndRunStep('encrypted', video, onEncrypted, test);
                    waitForEventAndRunStep('waitingforkey', video, onWaitingForKey, test);
                    waitForEventAndRunStep('playing', video, onPlaying, test);
                    video.src = '../content/test-encrypted.webm';
                    return video.setMediaKeys(mediaKeys);
                }).then(function(result) {
                    video.play();
                }).catch(function(error) {
                    forceTestFailureFromPromise(test, error);
                });
            }, 'Waiting for a key.');
        </script>
    </body>
</html>
